requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
cloudinary==1.41.0
crispy-bootstrap5==2024.2
dj-database-url==2.2.0
Django==5.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.1
django-crispy-forms==2.2
django-email==0.1.10
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
Markdown==3.7
MouseInfo==0.1.3
numpy==1.26.4
opencv-contrib-python==4.9.0.80
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==2.9.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.2.0
requests==2.32.3
setuptools==73.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wheel==0.44.0
whitenoise==6.7.0
